+++ library/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add `blob` and `blobAsync` schema
+
 ## v0.1.0 (July 12, 2023)
 
 - Initial release